+++ src/Addons/34thPRC_Overrides/data/Bulldogs/config_vehicles.hpp
@@ -1,0 +1,9 @@
+class Car_F;
+class DMNS_Bulldog_base : Car_F // DMNS_Vehicles_Bulldog
+{
+	enableRadio=1;
+	tf_range=25000;
+	tf_isolatedAmount=0.4;
+	tf_dialogUpdate="call TFAR_fnc_updateLRDialogToChannel;";
+	tf_hasLRradio=1;
+}